Split language controller into single APIs
