+++ src/Umbraco.Cms.ManagementApi/Controllers/Language/LanguageControllerBase.cs
@@ -7,6 +7,6 @@ namespace Umbraco.Cms.ManagementApi.Controllers.Language;
 [ApiController]
 [BackOfficeRoute("api/v{version:apiVersion}/language")]
 [OpenApiTag("Language")]
-public class LanguageControllerBase
+public class LanguageControllerBase : Controller
 {
 }
